Silence MSVC truncation warning in try_find_insert_bucket().
MSVC warns that truncation could occur in try_find_insert_bucket() because a
size_t (potentially 64-bit) is being assigned to an int:

cuckoohash_map.hh(1346): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data

The warning is benign, since slot_per_bucket is typically tiny. This commit
switches the loop variable to an int to silence that warning.

+++ src/cuckoohash_map.hh
@@ -1328,7 +1328,7 @@ private:
         (void)partial;
         slot = -1;
         bool found_empty = false;
-        for (size_t i = 0; i < slot_per_bucket; ++i) {
+        for (int i = 0; i < static_cast<int>(slot_per_bucket); ++i) {
             if (b.occupied(i)) {
                 if (!is_simple && partial != b.partial(i)) {
                     continue;
